requirements.txt
# ============================================================
# Log Sentinel v2.0 - Dependências
# ============================================================
# Autor: Duarte Cunha (Nº 2024271)
# ISTEC - Instituto Superior de Tecnologias Avançadas de Lisboa
# Ano Letivo: 2025/2026
# ============================================================
# Instalação: pip install -r requirements.txt
# ============================================================

# === GUI Framework ===
customtkinter>=5.2.0
pillow>=10.0.0

# === Machine Learning ===
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# === Visualização e Gráficos ===
matplotlib>=3.7.0

# === Exportação de Relatórios ===
reportlab>=4.0.0          # PDF generation
python-docx>=1.0.0        # Word documents
openpyxl>=3.1.0           # Excel files

# === Notificações Desktop ===
plyer>=2.1.0              # Cross-platform notifications

# === Processamento de Dados ===
python-dateutil>=2.8.0

# === Monitorização em Tempo Real (NOVO) ===
watchdog>=3.0.0           # File system monitoring

# === Instalador ===
pyinstaller>=6.0.0        # Para criar executável

# === Desenvolvimento (opcional) ===
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
